utils: Prevent multiple calls to 'customElements.define'
Currently, when the bookmarklet is clicked multiple times (without
reloading the page), the following error is thrown:

  Uncaught DOMException: CustomElementRegistry.define: 'x-bookmarklet'
  has already been defined as a custom element

Root Cause:

Each click on the bookmarklet creates a new instance of the
'BookmarkletRenderer' class. In its constructor, the static method
'WebComponent.defineReactComponent' is called, which internally invokes
'customElements.define'. However, there is no check to ensure that the
custom element 'x-bookmarklet' hasn't already been defined, leading to
the error.

Solution:

Add a check using 'customElements.get' before calling
'customElements.define'. If 'customElements.get' returns 'undefined',
it is safe to proceed with defining the custom element. This prevents
duplicate definitions and resolves the error.

Signed-off-by: Stefan Kühnel <git@stefankuehnel.com>